Moving prisma calls to lib\services
